+++ online-banking-system-backend/src/main/java/com/app/pojos/Account.java
@@ -1,3 +1,4 @@
+
 package com.app.pojos;
 
 import java.math.BigDecimal;
@@ -6,6 +7,9 @@ import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
@@ -22,16 +26,17 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Data
-public class Account extends BaseEntity {
+public class Account {
 
-	@Column(name = "account_no", nullable = false)
-	@Length(min = 15, max = 15)
-	private String accountNo;
-	
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column(name = "account_no")
+	private long accountNo;
+
 	@Column(nullable = false)
 	@Length(min = 4, max = 4)
 	private String pin;
-	
+
 	@Column(nullable = false)
 	@Range(min = 0, message = "Balance cannot be negative")
 	private BigDecimal balance;
@@ -39,9 +44,19 @@ public class Account extends BaseEntity {
 	@Enumerated(EnumType.STRING)
 	@Column(name = "account_type", nullable = false, length = 20)
 	private AccountType accountType;
-	
+
 	@ManyToOne
-	@JoinColumn(name = "user_id")
+	@JoinColumn(name = "customer_id")
+	// @Column(nullable = false)
 	private User user;
+
+	public Account(String pin, BigDecimal balance, AccountType accountType, User user) {
+		super();
+		this.pin = pin;
+		this.balance = balance;
+		this.accountType = accountType;
+		this.user = user;
+	}
 	
+
 }
