+++ online-banking-system-backend/src/main/java/com/app/pojos/Transaction.java
@@ -1,7 +1,9 @@
 package com.app.pojos;
 
-import java.time.LocalDateTime;
+import java.math.BigDecimal;
+import java.time.LocalDate;
 
+import javax.persistence.AttributeOverride;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
@@ -23,27 +25,25 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Data
+@AttributeOverride(column = @Column(name = "transaction_id"), name = "id")
 public class Transaction extends BaseEntity {
-
-	@Column(name = "transaction_from")
-	private String transactionFrom;
-	
-	@Column(name = "transaction_to")
-	private String transactionTo;
-	
-	@Range(min = 0)
-	private Long amount;
-	
-	@Past
-	@DateTimeFormat(pattern = "yyyy-MM-dd")
-	private LocalDateTime timestamp;
-	
-	@Enumerated(EnumType.STRING)
-	@Column(name = "transaction_status", length = 20)
-	private TransactionStatus transactionStatus;
-	
-	@ManyToOne
-	@JoinColumn(name = "user_id")
-	private User user;
-	
+    
+    @Column(name = "transaction_to")
+    private String transactionTo;
+    
+    @Range(min = 0)
+    private BigDecimal amount;
+    
+    @Past
+    @DateTimeFormat(pattern = "yyyy-MM-dd")
+    private LocalDate date;
+    
+    @Enumerated(EnumType.STRING)
+    @Column(name = "transaction_status", length = 20)
+    private TransactionStatus transactionStatus;
+    
+    @ManyToOne
+    @JoinColumn(name = "transaction_from")
+    private Account account;
+    
 }
