+++ online-banking-system-backend/src/main/resources/application.properties
@@ -6,7 +6,7 @@ server.port=8080
 spring.datasource.url=jdbc:mysql://localhost:3306/banking-system?createDatabaseIfNotExist=true&useSSL=false&allowPublicKeyRetrieval=true
 spring.datasource.username=root
 
-spring.datasource.password=root
+spring.datasource.password=Sayali@20
 
 # JPA properties
 spring.jpa.show-sql = true
