--- src/main/java/com/app/OnlineBankingSystemApplication.java
@@ -1,13 +1,0 @@
-package com.app;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class OnlineBankingSystemApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(OnlineBankingSystemApplication.class, args);
-	}
- 
-}
