--- src/main/java/com/app/pojos/AccountType.java
@@ -1,5 +1,0 @@
-package com.app.pojos;
-
-public enum AccountType {
-	SAVINGS, CURRENT, SALARY, LOAN
-}
