--- src/main/java/com/app/pojos/BaseEntity.java
@@ -1,16 +1,0 @@
-package com.app.pojos;
-
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.MappedSuperclass;
-
-import lombok.Data;
-
-@MappedSuperclass
-@Data
-public class BaseEntity {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-}
