--- src/main/java/com/app/pojos/Report.java
@@ -1,5 +1,0 @@
-package com.app.pojos;
-
-public class Report {
-
-}
