--- src/main/java/com/app/pojos/TransactionStatus.java
@@ -1,5 +1,0 @@
-package com.app.pojos;
-
-public enum TransactionStatus {
-	SUCCESS, FAILURE, PENDING
-}
