--- src/main/resources/application.properties
@@ -1,20 +1,0 @@
-#Tomcat server default port number 8080
-server.port=8080
-#context path default value if not given : empty
-#server.servlet.context-path=/spring-mvc-boot
-#DB properties
-spring.datasource.url=jdbc:mysql://localhost:3306/banking-system?createDatabaseIfNotExist=true&useSSL=false&allowPublicKeyRetrieval=true
-spring.datasource.username=root
-
-spring.datasource.password=Sayali@20
-
-# JPA properties
-spring.jpa.show-sql = true
-spring.jpa.hibernate.ddl-auto = update
-#spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL5Dialect
-#logging.level.org.springframework.orm.hibernate5=DEBUG
-# Spring MVC ViewResolver related
-#spring.mvc.view.prefix=/WEB-INF/views
-#spring.mvc.view.suffix=.jsp
-#Disable open session in view 
-spring.jpa.open-in-view=false
