--- src/test/java/com/app/OnlineBankingSystemApplicationTests.java
@@ -1,13 +1,0 @@
-package com.app;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class OnlineBankingSystemApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
